refactor: 💡 Code refacotring
✅ Closes: #140
