+++ src/grid/Cells/index.tsx
@@ -2,10 +2,9 @@ import React, { memo, useMemo } from "react";
 
 import { KonvaGroup } from "../../@konva";
 import { useTimelineContext } from "../../timeline/TimelineContext";
+import { dayDetail, timeBlockTz } from "../../utils/timeBlockArray";
 import GridCell from "../Cell";
 import GridCellGroup from "../CellGroup";
-
-import { dayDetail, timeBlockTz } from "./utils";
 
 interface GridCellsProps {
   height: number;
@@ -18,14 +17,17 @@ const GridCells = ({ height }: GridCellsProps) => {
     visibleTimeBlocks,
     resolution: { unitAbove },
   } = useTimelineContext();
+
   const tz = useMemo(() => interval.start!.toISO()!.slice(-6), [interval]);
 
   const dayInfo = useMemo(
     () => dayDetail(unitAbove, aboveTimeBlocks, interval),
     [unitAbove, aboveTimeBlocks, interval]
   );
+
   const aboveHourInfo = useMemo(() => timeBlockTz(aboveTimeBlocks, tz), [tz, aboveTimeBlocks]);
   const visibileHourInfo = useMemo(() => timeBlockTz(visibleTimeBlocks, tz), [tz, visibleTimeBlocks]);
+
   return (
     <KonvaGroup>
       {aboveTimeBlocks.map((column, index) => (
