+++ src/utils/timeBlockArray.ts
@@ -1,6 +1,6 @@
 import { Interval } from "luxon";
 
-import { daysInMonth, getMonth, getStartMonthsDay, getYear, Scale } from "../../utils/time-resolution";
+import { daysInMonth, getMonth, getStartMonthsDay, getYear, Scale } from "./time-resolution";
 
 interface VisibleHourInfoProps {
   backHour?: boolean;
@@ -14,6 +14,7 @@ interface DayDetailProps {
 
 export const timeBlockTz = (timeBlock: Interval[], initialTz?: string) => {
   const dayInfoArray: VisibleHourInfoProps[] = [];
+
   timeBlock.forEach((column) => {
     const tzStart = column.start!.toISO()?.slice(-6);
 
@@ -23,29 +24,36 @@ export const timeBlockTz = (timeBlock: Interval[], initialTz?: string) => {
           backHour: true,
           nextHour: false,
         });
+
         return;
       }
+
       dayInfoArray.push({
         backHour: false,
         nextHour: true,
       });
     }
+
     dayInfoArray.push({
       backHour: false,
       nextHour: false,
     });
+
     return;
   });
+
   return dayInfoArray;
 };
 
 export const dayDetail = (unitAbove: Scale, aboveTimeBlocks: Interval[], interval: Interval) => {
   if (unitAbove === "month") {
     const dayInfo: DayDetailProps[] = [];
+
     aboveTimeBlocks.forEach((column, index) => {
       const month = getMonth(column);
       const year = getYear(column);
       const currentMonthDays = daysInMonth(Number(month), Number(year));
+
       if (index === 0) {
         const startDay = getStartMonthsDay(interval.start!);
         const daysToMonthEnd = currentMonthDays - Number(startDay) + 1;
@@ -53,15 +61,17 @@ export const dayDetail = (unitAbove: Scale, aboveTimeBlocks: Interval[], interva
           thisMonth: daysToMonthEnd,
           untilNow: daysToMonthEnd,
         });
+
         return;
       }
-      const n = dayInfo[index - 1].untilNow! + currentMonthDays;
 
+      const n = dayInfo[index - 1].untilNow! + currentMonthDays;
       dayInfo.push({
         thisMonth: currentMonthDays,
         untilNow: n,
       });
     });
+
     return dayInfo;
   }
   return [];
